+++ src/main/java/bots/CalendarBot/CalendarBirthday.java
@@ -48,7 +48,7 @@ public class CalendarBirthday extends Calendar {
 				insert(new EventBirthday(value[1], Integer.parseInt(value[0]), Integer.parseInt(value[2]), Integer.parseInt(value[3]), firstYear, defaultTumtum));
 			}
 		} catch (Exception e) {
-			
+			CalendarBot.printMessage(ErrorMessages.eventsNotLoadError(this), false);
 		}		
 	}
 }